requirements.txt
fastapi>=0.112.0,<0.113.0
uvicorn[standard]
httpx==0.28.1
pydantic>=2.0,<3.0
python-multipart
python-dotenv
supabase
anthropic
langgraph == 0.6.6
langchain == 0.3.27
langchain-anthropic == 0.3.19
